Implement ObjectJsonWriter.
https://github.com/square/moshi/issues/89
